LayoutTests/sputnik/Conformance/12_Statement/12.6_Iteration_Statements/12.6.2_The_while_statement/S12.6.2_A13_T3-expected.txt

CONSOLE MESSAGE: line 76: SyntaxError: Unexpected token '('
S12.6.2_A13_T3

PASS Expected parsing failure

TEST COMPLETE
